pyproject.toml
[project]
name = "upkeep"
version = "1.0.0"
description = "Django Application"
readme = "README.md"
requires-python = ">=3.12"
classifiers = ["Private :: Do Not Upload"]
dependencies = [
    "crispy-bootstrap5>=2024.10",
    "django-crispy-forms>=2.3",
    "django-htmx>=1.21.0",
    "django>=5.1.3",
    "environs[django]>=11.0.0",
    "gunicorn>=23.0.0 ; sys_platform == 'linux'",
    "psycopg[binary,pool]>=3.2.3 ; sys_platform == 'linux'",
    "whitenoise>=6.8.2",
    "python-dateutil>=2.9.0.post0",
]

[tool.hatch.build.targets.wheel]
include = ["/config", "/upkeep"]

[project.scripts]
upkeep-manage = "config.manage:main"

[dependency-groups]
dev = [
    { include-group = "test" },
    { include-group = "typing" },
    { include-group = "lint" },
    "django-debug-toolbar>=4.4.6",
    "ipython>=8.29.0",
    "rich >= 13.9.4",
]
typing = [
    { include-group = "test" },
    "django-stubs[compatible-mypy]>=5.1.1",
    "mypy>=1.13.0",
    "types-python-dateutil>=2.9.0.20241003",
]
test = [
    "pytest-cov>=6.0.0",
    "pytest-django>=4.9.0",
    "pytest>=8.3.3",
]
lint = [
    "djade>=1.3.2",
    "ruff>=0.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"

[tool.coverage.run]
omit = ["manage.py", "wsgi.py", "asgi.py", "__init__.py"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = "crispy_forms.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "crispy_bootstrap5.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.ruff]
line-length = 99

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]
